Event handlers are now inner classes within EventManagerRTG.
Doing this has also resolved the decoration issue, but further testing
still required.
