--- src/main/java/rtg/event/bus/oregen/oregenevent/GenerateMinableRTG.java
@@ -1,73 +1,0 @@
-package rtg.event.bus.oregen.oregenevent;
-
-import net.minecraftforge.event.terraingen.OreGenEvent;
-import rtg.config.rtg.ConfigRTG;
-import cpw.mods.fml.common.eventhandler.Event.Result;
-import cpw.mods.fml.common.eventhandler.SubscribeEvent;
-
-public class GenerateMinableRTG
-{
-
-    public GenerateMinableRTG()
-    {
-
-    }
-
-    @SubscribeEvent
-    public void onGenerateMinable(OreGenEvent.GenerateMinable event) {
-
-        switch (event.type) {
-            
-            case COAL:
-                
-                if (!ConfigRTG.generateOreCoal) {
-                    event.setResult(Result.DENY);
-                }
-                
-                break;
-            
-            case IRON:
-                
-                if (!ConfigRTG.generateOreIron) {
-                    event.setResult(Result.DENY);
-                }
-                
-                break;
-            
-            case GOLD:
-                
-                if (!ConfigRTG.generateOreGold) {
-                    event.setResult(Result.DENY);
-                }
-                
-                break;
-            
-            case DIAMOND:
-                
-                if (!ConfigRTG.generateOreDiamond) {
-                    event.setResult(Result.DENY);
-                }
-                
-                break;
-            
-            case REDSTONE:
-                
-                if (!ConfigRTG.generateOreRedstone) {
-                    event.setResult(Result.DENY);
-                }
-                
-                break;
-            
-            case LAPIS:
-                
-                if (!ConfigRTG.generateOreLapis) {
-                    event.setResult(Result.DENY);
-                }
-                
-                break;
-            
-            default:
-                break;
-        }
-    }
-}
